// src/Enums/complaint.enum.ts
export enum complaintsType {
    VIOLATION_OF_CONTRACT_AGREEMENT = "violation_of_contract_agreement",
    FAILED_PAYMENT = "failed_payment",
    ORDERS_NOT_RECIEVED = "orders_not_recieved",
    ABUSE = "abuse",
    WALLET_RELATED_ISSUES = "wallet_related_issues",
    OTHER_ISSUES = "other_issues"
  }

export enum ComplaintResolutionLevel{
    RESOLVED = 'Resolved',
    UNRESOLVED = 'Unresolved',
    PROCESSING = 'Processing',
    

}